Remount USB stick RW to export print profile.

+++ src/printprofilemanager.cpp
@@ -1,3 +1,7 @@
+#include <cerrno>
+#include <cstring>
+#include <sys/mount.h>
+#include <QtCore>
 #include "pch.h"
 
 #include "profilesjsonparser.h"
@@ -60,11 +64,24 @@ bool PrintProfileManager::exportProfiles(const QString& mountPoint )
 {
     QString destPath { QString("%1/print-profiles.json").arg(mountPoint) };
     QFile sourceFile(PrintProfilesPath);
+    bool result;
 
     debug("+ PrintProfileManager::exportProfiles\n");
     debug("  + mountPoint: %s\n", mountPoint.toUtf8().data());
     debug("  + destPath: %s\n", destPath.toUtf8().data());
 
+    if (mount(nullptr, mountPoint.toUtf8().data(), nullptr, MS_REMOUNT, nullptr) != 0) {
+        debug("  + failed to remount %s rw: %s\n", mountPoint.toUtf8().data(), strerror(errno));
+        return false;
+    }
+
     QFile::remove(destPath);
-    return sourceFile.copy(destPath);
+    result = sourceFile.copy(destPath);
+
+    if (mount(nullptr, mountPoint.toUtf8().data(), nullptr, MS_REMOUNT | MS_RDONLY, nullptr) != 0) {
+        debug("  + failed to remount %s ro: %s\n", mountPoint.toUtf8().data(), strerror(errno));
+        return false;
+    }
+
+    return result;
 }
